Targets .NET 10 (#3353)
* Updates projects to target .NET 10

Updates the target framework of several projects to .NET 10.
This ensures compatibility and takes advantage of the latest .NET features.

Updates package versions to align with the target framework update.

* Updates Maui controls version for .NET 10

Updates the Microsoft.Maui.Controls package version to 10.0.10 for .NET 10 targets in Maui projects.

This ensures that the projects use the correct Maui version when targeting .NET 10.
It also temporarily removes conditional package versioning from Directory.Packages.props which is now handled within the project files themselves.

* Updates Uno check version and target platforms

Updates the Uno check version to 1.33.1 in both the build and CI workflows.

Adds tvos as a target platform for workload installation in both workflows to support tvOS builds.

* Fixes build targets

Also updates target frameworks to include net10.

* Adds OutputType property to project files

Specifies the OutputType property as Library in the project files.

This ensures that the projects are built as libraries, which is necessary for proper packaging and referencing in other projects.

* Enhances XAML Navigation and Dialog Features

Adds attached properties for tabbed page navigation and appearance.
Introduces the ability to specify navigation types when going back.
Adds logging for navigation failures and provides a way to add services
required by markup extensions to prevent runtime errors.

* Disables code signing in CI workflows

Turns off code signing for .NET 10 builds, likely for testing or development purposes where signed builds are not immediately required.

* Registers ILoggerFactory for Navigation Tests

Registers an instance of ILoggerFactory in the test container
for the Navigation Tests. This enables logging functionality
within the tests, enhancing debugging and diagnostics.